IntelligentLogInsightsAPI/function_app/requirements.txt
azure-functions
azure-cosmos
azure-servicebus==7.8.1
azure-cosmos==4.4.0
